+++ src/router/routes.js
@@ -1,0 +1,4 @@
+export const vueRoutes = Object.freeze({
+  rolesAndRules: { name: 'roles-and-rules' },
+  rolesAndRulesIndex: { name: 'roles-and-rules.index' },
+})