+++ src/router/routes/UserRoutes.js
@@ -1,6 +1,7 @@
 // TODO refactor, extract require
 
 import { authorizedGuard } from '../helpers/navigationGuards'
+import { vueRoutes } from '../routes'
 
 // Each of these routes are loaded asynchronously,
 // when a user first navigates to each corresponding endpoint.
@@ -414,6 +415,24 @@ export const UserRoutes = {
           name: 'keyValue.index',
           component: function (resolve) {
             require(['../../components/User/KeyValue/KeyValue.Index.vue'], resolve)
+          },
+        },
+      ],
+    },
+
+    {
+      path: '/roles-and-rules',
+      name: vueRoutes.rolesAndRules.name,
+      component: function (resolve) {
+        require(['../../components/User/RolesAndRules/RolesAndRules.vue'], resolve)
+      },
+      redirect: vueRoutes.rolesAndRulesIndex,
+      children: [
+        {
+          path: '',
+          name: vueRoutes.rolesAndRulesIndex.name,
+          component: function (resolve) {
+            require(['../../components/User/RolesAndRules/RolesAndRules.index.vue'], resolve)
           },
         },
       ],