+++ src/validators.js
@@ -1,5 +1,6 @@
 import * as validators from 'vuelidate/lib/validators'
 import { base } from '@tokend/js-sdk'
+import { ADMIN_CONST } from '@/constants'
 
 export const password = value => validators.minLength(8)(value)
 export const seed = value => base.Keypair.isValidSecretKey(value)
@@ -13,6 +14,14 @@ export const emailOrAccountId = value => {
 export const emailOrAccountIdOrBalanceId = value => {
   return validators.email(value) || accountId(value) || balanceId(value)
 }
+
+export const isNotAdminRule = (value) => +value !== ADMIN_CONST.RULE_ID
+
+export const isRuleNotAdded = (rulesID) =>
+  (value) => rulesID.indexOf(value) === -1
+
+export const isRuleExists = (rulesID) =>
+  (value) => rulesID.indexOf(value) !== -1
 
 export const noMoreThanAvailableForIssuance = available => value => {
   return +available >= +value